+++ numba/cuda/device_init.py
@@ -2,12 +2,11 @@
 from .stubs import (threadIdx, blockIdx, blockDim, gridDim, laneid,
                     warpsize, syncthreads, syncthreads_count, syncwarp,
                     syncthreads_and, syncthreads_or, shared, local,
-                    const, gridsize, atomic, shfl_sync_intrinsic,
-                    vote_sync_intrinsic, match_any_sync, match_all_sync,
-                    threadfence_block, threadfence_system,
-                    threadfence, selp, popc, brev, clz, ffs, fma, cbrt,
-                    cg, activemask, lanemask_lt, nanosleep)
-from .intrinsics import grid
+                    const, atomic, shfl_sync_intrinsic, vote_sync_intrinsic,
+                    match_any_sync, match_all_sync, threadfence_block,
+                    threadfence_system, threadfence, selp, popc, brev, clz,
+                    ffs, fma, cbrt, cg, activemask, lanemask_lt, nanosleep)
+from .intrinsics import grid, gridsize
 from .cudadrv.error import CudaSupportError
 from numba.cuda.cudadrv.driver import (BaseCUDAMemoryManager,
                                        HostOnlyCUDAMemoryManager,
